refactor: split account filters
